Add test outputs
Outputs are the same as the original FORTRAN version except for NEWUOA
where there are some (small) differences. `*-test.out` is the output
with the current version of the code. `*-test.out~orig` is the output
with the original FORTAN version (omitted if same as current version),
`*-test.out~old` is the output from an older C version, and
`*-test.out~bigfloat` is the output with trigonometric tables computed
with "BigFloat" precision (which results in slightly more calls to the
objective function, this version of the table is thus disabled by
default).